machine: use native WriteRegister/ReadRegister functions on mimxrt1062
This seems better than adding extra WriteRegisterEx/ReadRegisterEx
methods that do the same thing but use hardware features instead of
implementing it in software.
