+++ src/machine/i2c.go
@@ -15,6 +15,8 @@ var _ interface { // 2
 	Configure(config I2CConfig) error
 	Tx(addr uint16, w, r []byte) error
 	SetBaudRate(br uint32) error
+	WriteRegister(address, register uint8, data []byte) error
+	ReadRegister(address, register uint8, data []byte) error
 } = (*I2C)(nil)
 
 // TWI_FREQ is the I2C bus speed. Normally either 100 kHz, or 400 kHz for high-speed bus.
@@ -67,26 +69,3 @@ const (
 	// I2CModeTarget represents an I2C peripheral in target mode.
 	I2CModeTarget
 )
-
-// WriteRegister transmits first the register and then the data to the
-// peripheral device.
-//
-// Many I2C-compatible devices are organized in terms of registers. This method
-// is a shortcut to easily write to such registers. Also, it only works for
-// devices with 7-bit addresses, which is the vast majority.
-func (i2c *I2C) WriteRegister(address uint8, register uint8, data []byte) error {
-	buf := make([]uint8, len(data)+1)
-	buf[0] = register
-	copy(buf[1:], data)
-	return i2c.Tx(uint16(address), buf, nil)
-}
-
-// ReadRegister transmits the register, restarts the connection as a read
-// operation, and reads the response.
-//
-// Many I2C-compatible devices are organized in terms of registers. This method
-// is a shortcut to easily read such registers. Also, it only works for devices
-// with 7-bit addresses, which is the vast majority.
-func (i2c *I2C) ReadRegister(address uint8, register uint8, data []byte) error {
-	return i2c.Tx(uint16(address), []byte{register}, data)
-}
